Extract commands to own module
